fix: user not being able to log in using OAuth and basic auth at the same time
If the user had registered using both the OAuth and the basic (local) authentication, he could not log in because the loadUserByUsername method in the ImplementsUserDetailsService would query by filtering only by the email address, wich would cause to receive two users instead of one, breaking the system. Having said that, changing the quering method from findByEmail to checkIfUserExists, which checks not only the email, but also the provider, solved the issue.